Refactor music-on-hold
