+++ packages/voice/src/models/v1/manage-conference-participant-request/manage-conference-participant-request.ts
@@ -1,4 +1,4 @@
-import { MusicOnHold } from '../music-on-hold';
+import { MusicOnHold } from '../enums';
 
 /**
  *
