+++ packages/voice/src/models/v1/mod-svaml/svaml-action-connect-sip/svaml-action-connect-sip.ts
@@ -1,5 +1,5 @@
 import { CallHeader } from '../../call-header';
-import { MusicOnHold } from '../../music-on-hold';
+import { MusicOnHold } from '../../enums';
 
 /**
  * Determines how to route a call to a SIP server. Available to use in a response to an [Incoming Call Event](../../../voice/tag/Callbacks/#tag/Callbacks/operation/ice) callback.
@@ -15,7 +15,7 @@ export interface SvamlActionConnectSip {
   /** Used to override the CLI (or caller ID) of the client. The phone number of the person who initiated the call is shown as the CLI. To set your own CLI, you may use your verified number or your Dashboard virtual number. */
   cli?: string;
   /** An optional parameter to specify the SIP transport protocol. If unspecified, UDP is used. */
-  transport?: Transport;
+  transport?: 'udp' | 'tcp' | 'tls';
   /** If enabled, suppresses [ACE](../../voice/tag/Callbacks/#tag/Callbacks/operation/ace) and [DICE](../../voice/tag/Callbacks/#tag/Callbacks/operation/dice) callbacks for the call. */
   suppressCallbacks?: boolean;
   /** [Private SIP headers](../../sip-trunking/#receiving-calls-from-sinch-platform-to-your-sip-infrastructure) to send with the call. */
@@ -23,8 +23,6 @@ export interface SvamlActionConnectSip {
   /** Means \"music on hold\". If this optional parameter is included, plays music to the connected participant if the SIP call is placed on hold. If `moh` isn't specified and the SIP call is placed on hold, the user will only hear silence while during the holding period . */
   moh?: MusicOnHold;
 }
-
-type Transport = 'udp' | 'tcp' | 'tls';
 
 export interface SipDestination {
