--- packages/voice/src/models/v1/music-on-hold/index.ts
@@ -1,1 +1,0 @@
-export type { MusicOnHold } from './music-on-hold';
