--- packages/voice/src/models/v1/music-on-hold/music-on-hold.ts
@@ -1,1 +1,0 @@
-export type MusicOnHold = 'ring' | 'music1' | 'music2' | 'music3';
